fusion_workflow/requirements.txt
# Fusion Workflow Requirements

# Python 3.11+
# python-dotenv - 环境变量管理
python-dotenv>=1.0.0

# OpenAI SDK - 用于LLM调用（V2学习系统依赖）
openai>=1.0.0

# asyncio - Python原生，无需安装
# aiohttp - 异步HTTP（OpenAI SDK依赖）
# pydantic - 数据验证（OpenAI SDK依赖）

# 可选依赖
# portalocker - 文件锁（后续阶段）
# pytest - 测试框架（后续阶段）
